requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-surprise==1.1.4
streamlit==1.40.1
joblib==1.4.2
annoy==1.17.3